run-container: support connecting to koji via FQDN
Support connecting to koji via its fully qualified domain name, i.e.
org.osbuild.koji.koji. For this an extra SSL certificate is needed,
that has the FQDN as the subject, as well as a kerberos principal.

NB: This needs to have a apache VirtualHost section for the FQDN
that uses the new certificate as well.