koji-osbuild.spec: don't generate from .in file
Remove koji-osbuild.spec.in and the bits in meson.build that generate
the final spec file out of it. Pull in the changelog from how Ondřej
Budai does it in osbuild-composer, because it's adorable.

When releasing, the version now has to be bumped in the spec file.

This is to make it consistent with other osbuild projects, and to
simplify reverse dependency testing.