test_koji.sh: refactor the test case and de-duplicate some code
Refactor the test case implementation to make it easier to reuse
existing code when adding a new testing scenario.

The common pieces were moved under a new `SutInfo` class, which provides
information and convenience methods related to the system on which the
test is running.

This change will make it easier to later add a new scenario testing
upload to the cloud.