hub: mark `baseurl` as required in `repo` schema
Previously it the `repo` schema didn't mark the `baseurl` property as
required, although Cloud API "technically" requires it and the builder
code also made assumptions about it being always provided.

Note on Cloud API:
Cloud API requires that at least one of `baseurl`, `metalink` or
`mirrorlist` is specified for a repo. Since the plugin schema specifies
only `baseurl` as a property, it is as a result mandatory in order for
the Cloud API request to be valid and succeed.

Mark `baseurl` property as required in the `repo` schema and add unit
test for it.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>